ExampleLevel: Set up meshes for Berry plant
There are two pickups for berries in the level, and neither of them is
visible or usable. Set up meshes for one plant so that it can be used.

There is a second berry pickup used to group all the plant pickups. As
this makes adjusting its position extremely annoying, simply remove it
and put the plant pickups in a subfolder.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>