ExampleLevel: Add Somersloop and Mercer Spheres
Since Somersloops and Mercer Spheres are used in 1.0 for some things,
add some pickups in ExampleLevel. Add them as magic plant pickups and
as shrines (as found in base game). Some things are still a bit rough
around the edges, though: the shrines don't properly respond when the
pickup is collected. But the maddening voices can be heard. Don't let
them *CONSUME* your sanity, though...

Signed-off-by: Angel Pons <th3fanbus@gmail.com>