ExampleLevel: Make dynamic resource nodes pretty
The Blueprint spaghetti that handled dynamic resource nodes did not
account for mesh scale and offset, and Crude Oil nodes had problems
with the decal being too small and too tall. To make matters worse,
SAM nodes used the Limestone mesh with a pre-1.0 material.

Instead of making the spaghetti even worse, rip it all out and make
use of a map to associate resource descriptors with mesh parameters
within an ad-hoc struct type. Crude Oil is still handled separately
as it uses a decal. Also, ensure all nodes are about the same size,
and that the mesh is centered on the resource node's origin.

ExampleLevel also tried spawning resource nodes for resources added
by other mods, but had to use a fallback mesh as there is no way to
know which mesh or materials the modded resource's nodes would use.
Deal with this by using the resource descriptor's item mesh, scaled
up to about the size of a resource node.

Thanks to AngryBeaver for the suggestion of using a map for this.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>